client/requirements_serial.txt
pyserial==3.5
matplotlib==3.10.*